src/phare/include.h -> src/phare/phare.h
RuntimeDiagnosticInterface moved to simulator.h and renamed
SimulatorDiagnostics
